refactor: handtracking, webcam 폴더 구분
