+++ src/features/handtracking/solveProblem.ts
